requirements.txt
google-auth
google-cloud-compute
google-cloud-core
google-cloud-storage
python-dateutil
